# Conflicts:
#	Tests/Linq/DataProvider/OracleTests.cs
#	Tests/Linq/Update/UpdateTests.cs

+++ Source/LinqToDB/DataProvider/Oracle/OracleSqlBuilder.cs
@@ -278,13 +278,6 @@ namespace LinqToDB.DataProvider.Oracle
 					if (_identityField != null)
 						return 3;
 					break;
-
-				case SqlDropTableStatement dropTable:
-					_identityField = dropTable.Table.Fields.Values.FirstOrDefault(f => f.IsIdentity);
-					if (_identityField != null)
-						return 3;
-
-					break;
 			}
 
 			return base.CommandCount(statement);
@@ -292,7 +285,9 @@ namespace LinqToDB.DataProvider.Oracle
 
 		protected override void BuildDropTableStatement(SqlDropTableStatement dropTable)
 		{
-			if (_identityField == null)
+			var identityField = dropTable.Table.Fields.Values.FirstOrDefault(f => f.IsIdentity);
+
+			if (identityField == null && dropTable.IfExists == false)
 			{
 				base.BuildDropTableStatement(dropTable);
 			}
@@ -303,11 +298,52 @@ namespace LinqToDB.DataProvider.Oracle
 					: dropTable.Table.Schema + ".";
 
 				StringBuilder
-					.Append("DROP TRIGGER ")
-					.Append(schemaPrefix)
-					.Append("TIDENTITY_")
-					.Append(dropTable.Table.PhysicalName)
-					.AppendLine();
+					.AppendLine(@"BEGIN");
+
+				if (identityField != null)
+					StringBuilder
+						.Append("\tEXECUTE IMMEDIATE 'DROP TRIGGER ")
+						.Append(schemaPrefix)
+						.Append("TIDENTITY_")
+						.Append(dropTable.Table.PhysicalName)
+						.AppendLine("';")
+						.Append("\tEXECUTE IMMEDIATE 'DROP SEQUENCE ")
+						.Append(schemaPrefix)
+						.Append("SIDENTITY_")
+						.Append(dropTable.Table.PhysicalName)
+						.AppendLine("';")
+						;
+
+				StringBuilder
+					.Append("\tEXECUTE IMMEDIATE 'DROP TABLE ");
+				BuildPhysicalTable(dropTable.Table, null);
+				StringBuilder
+					.AppendLine("';")
+					;
+
+				if (dropTable.IfExists)
+				{
+					StringBuilder
+						.AppendLine("EXCEPTION")
+						.AppendLine("\tWHEN OTHERS THEN")
+						.Append("\t\tIF SQLCODE != -942 ")
+						;
+
+					if (identityField != null)
+						StringBuilder
+							.Append("AND SQLCODE != -4080 AND SQLCODE != -2289 ")
+							;
+
+					StringBuilder
+						.AppendLine("THEN")
+						.AppendLine("\t\t\tRAISE;")
+						.AppendLine("\t\tEND IF;")
+						;
+				}
+
+				StringBuilder
+					.AppendLine("END;")
+					;
 			}
 		}
 
@@ -344,24 +380,6 @@ END;",
 						;
 
 					break;
-
-				case SqlDropTableStatement dropTable:
-					if (commandNumber == 1)
-					{
-						StringBuilder
-							.Append("DROP SEQUENCE ")
-							.Append(GetSchemaPrefix(dropTable.Table))
-							.Append("SIDENTITY_")
-							.Append(dropTable.Table.PhysicalName)
-							.AppendLine();
-					}
-					else
-					{
-						base.BuildDropTableStatement(dropTable);
-					}
-
-					break;
-
 				case SqlCreateTableStatement createTable:
 				{
 					var schemaPrefix = GetSchemaPrefix(createTable.Table);
